@@ -8,7 +8,7 @@
     "dist"
   ],
   "dependencies": {
-    "rxjs": "^6.0.0"
+    "rxjs": "^7.8.0"
   },
   "devDependencies": {
     "@types/chai": "4.1.2",
@@ -17,12 +17,12 @@
     "del-cli": "^1.0.0",
     "mocha": "5.0.0",
     "source-map-support": "^0.5.9",
-    "ts-node": "4.1.0",
+    "ts-node": "10.9.1",
     "tslint": "^5.11.0",
-    "typescript": "2.9.2"
+    "typescript": "5.0.4"
   },
   "peerDependencies": {
-    "rxjs": "^6.0.0"
+    "rxjs": "^7.8.0"
   },
   "scripts": {
     "clean": "del-cli dist",
