@@ -1,5 +1,5 @@
 import {assert} from "chai";
-import {of, timer} from "rxjs";
+import { firstValueFrom, of, timer } from "rxjs";
 import {take} from "rxjs/operators";
 import {input} from "./InputState";
 
@@ -96,7 +96,7 @@ describe("InputState", function () {
 
     it("putFromPromise", function (done) {
         const s1 = input(0);
-        s1.clearAndPutFromPromise(timer(0).pipe(take(1)).toPromise());
+        s1.clearAndPutFromPromise(firstValueFrom(timer(0)));
         assert.isFalse(s1.hasValue());
         s1.values$().subscribe(val => {
             assert.equal(val, 0);
@@ -107,7 +107,7 @@ describe("InputState", function () {
 
     it("hasActivePromiseRequest", function (done) {
         const s1 = input(0);
-        s1.clearAndPutFromPromise(timer(0).pipe(take(1)).toPromise());
+        s1.clearAndPutFromPromise(firstValueFrom(timer(0)));
         assert.isFalse(s1.hasValue());
         assert.isTrue(s1.hasActivePromiseRequest());
 
@@ -122,7 +122,7 @@ describe("InputState", function () {
     it("isPristine", function () {
         const s1 = input<number>();
         assert.isTrue(s1.isPristine());
-        s1.clearAndPutFromPromise(timer(0).pipe(take(1)).toPromise());
+        s1.clearAndPutFromPromise(firstValueFrom(timer(0)));
         assert.isFalse(s1.isPristine());
         assert.isFalse(s1.hasValue());
     });
@@ -130,7 +130,7 @@ describe("InputState", function () {
     it("putFromPromiseIfPristine", function (done) {
         const s1 = input<number>();
 
-        s1.putFromPromiseIfPristine(() => timer(0).pipe(take(1)).toPromise());
+        s1.putFromPromiseIfPristine(() => firstValueFrom(timer(0)));
         assert.isTrue(s1.hasActivePromiseRequest());
         assert.isFalse(s1.hasValue());
 
