@@ -1,9 +1,9 @@
+export * from "./State";
 export * from "./Combiner";
 export * from "./StatesGroup";
 export * from "./DerivedState";
 export * from "./InputState";
 export * from "./MultiInputState";
-export * from "./State";
 export * from "./utils";
 export * from "./log";
 
