@@ -1,7 +1,7 @@
 import {assert} from "chai";
 import {input} from "./InputState";
 import {State} from "./State";
-import {IfThen} from "./utils";
+import {IfThen} from "./state-utils";
 
 describe("Utils", function () {
 
